Refactor model packages
