+++ src/main/java/seedu/address/logic/Logic.java
@@ -9,7 +9,7 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.assignment.Assignment;
-import seedu.address.model.assignment.Task;
+import seedu.address.model.task.Task;
 
 /**
  * API of the Logic component
