+++ src/main/java/seedu/address/model/Model.java
@@ -6,7 +6,7 @@ import java.util.function.Predicate;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.assignment.Assignment;
-import seedu.address.model.assignment.Task;
+import seedu.address.model.task.Task;
 import seedu.address.timetable.TimetableData;
 
 /**
