+++ src/main/java/seedu/address/model/assignment/NameContainsKeywordsPredicate.java
@@ -6,7 +6,7 @@ import java.util.function.Predicate;
 import seedu.address.commons.util.StringUtil;
 
 /**
- * Tests that a {@code Assignment}'s {@code Name} matches any of the keywords given.
+ * Tests that an {@code Assignment}'s {@code Name} matches any of the keywords given.
  */
 public class NameContainsKeywordsPredicate implements Predicate<Assignment> {
     private final List<String> keywords;
