+++ src/main/java/seedu/address/model/lesson/Lesson.java
@@ -4,10 +4,10 @@ import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
 import java.util.Objects;
 
-import seedu.address.model.assignment.Deadline;
-import seedu.address.model.assignment.ModuleCode;
-import seedu.address.model.assignment.Name;
-import seedu.address.model.assignment.Task;
+import seedu.address.model.task.Deadline;
+import seedu.address.model.task.ModuleCode;
+import seedu.address.model.task.Name;
+import seedu.address.model.task.Task;
 
 public class Lesson extends Task {
     private final Deadline endTime;
