+++ src/main/java/seedu/address/model/task/ModuleCode.java
@@ -1,4 +1,4 @@
-package seedu.address.model.assignment;
+package seedu.address.model.task;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
