+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -25,9 +25,9 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.EditCommand.EditAssignmentDescriptor;
-import seedu.address.model.assignment.Deadline;
-import seedu.address.model.assignment.ModuleCode;
-import seedu.address.model.assignment.Name;
+import seedu.address.model.task.Deadline;
+import seedu.address.model.task.ModuleCode;
+import seedu.address.model.task.Name;
 import seedu.address.testutil.EditAssignmentDescriptorBuilder;
 
 public class EditCommandParserTest {
