+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -18,9 +18,9 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import seedu.address.model.assignment.Assignment;
-import seedu.address.model.assignment.Task;
 import seedu.address.model.assignment.exceptions.DuplicateAssignmentException;
 import seedu.address.model.lesson.Lesson;
+import seedu.address.model.task.Task;
 import seedu.address.testutil.AssignmentBuilder;
 import seedu.address.timetable.TimetableData;
