+++ src/test/java/seedu/address/model/task/DeadlineTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.assignment;
+package seedu.address.model.task;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
