+++ src/test/java/seedu/address/testutil/LessonBuilder.java
@@ -1,9 +1,9 @@
 package seedu.address.testutil;
 
-import seedu.address.model.assignment.Deadline;
-import seedu.address.model.assignment.ModuleCode;
-import seedu.address.model.assignment.Name;
 import seedu.address.model.lesson.Lesson;
+import seedu.address.model.task.Deadline;
+import seedu.address.model.task.ModuleCode;
+import seedu.address.model.task.Name;
 
 /**
  * A utility class to help with building Lesson objects.
